Add an RGB matrix preview that triggers whenever the RGB_MOD key is pressed
Now that effects only show up on the base layer, when we try to change
effects, we don't know what we are changing to. Same with changing HSV
values. With this change, when pressing down the keys that make these
changes, the keyboard will show the effect, despite not being on the
base layer.