Disable SIDE_SYNC, as it breaks certain RGB effects in some cases
When we flash the two halves of the keyboard with different firmwares
(the only difference being that the left half has MASTER_LEFT and the
right has MASTER_RIGHT), keypress-based effects don't work properly.
Instead they mirror the master half onto the slave half. Turns out this
is caused by SIDE_SYNC. Although the hack that we do with SIDE_SYNC
(telling the slave that it is wrong about the side it is on) works with
the indicator LEDs that we set from code, it does not work for the
keypress-based effects.

I think this is because for those effects, the master uses its own
keycap to tell the slave which LEDs to light up, so the inversion caused
by the slave thinking it's on the opposite side ends up inverting these
LEDs. This doesn't happen with the indicator lights, because for those,
the slave determines which ones to light up based on its own already
inverted keymap, and then the inversion due to SIDE_SYNC fixes it.