Fix BSP_DEL key behavior when holding down the key
We are trying to change the behavior of the backspace key when in the
shifted state by writing custom code that runs when the key is
pressed/released. But we're currently not handling the case where the
shift state changes while the key is being held.

Possible ways to handle this case:
  1. have the shift key unregister all keys that are affected
     - would need to keep track of all such keys
     - would also want to handle this generically for all modifier keys
  2. instead of register/unregister, just do taps
     - then how do we respect the OS's configured key repeat speed?
  3. keep track of which one is pressed and on release, unregister it

Decided to go with option 3.